Reorganize data export methods into new service
