+++ src/AppServices/Utilities/DataExport.cs
@@ -4,17 +4,27 @@ namespace Cts.AppServices.Utilities;
 
 public static class DataExport
 {
+    /// <summary>
+    /// Creates an Excel spreadsheet from <see cref="IEnumerable{T}"/> records. 
+    /// </summary>
+    /// <param name="records">The records to add to the spreadsheet.</param>
+    /// <param name="sheetName">A name for the worksheet.</param>
+    /// <param name="deleteLastColumn">A flag indicating whether to delete the final column in the table.</param>
+    /// <typeparam name="T">The type of the records being inserted.</typeparam>
+    /// <returns>An Excel spreadsheet with a worksheet named <paramref name="sheetName"/> containing the data in
+    /// <paramref name="records"/> as a table.</returns>
     public static MemoryStream AsExcelStream<T>(this IEnumerable<T> records, string sheetName,
-        bool hideLastColumn = false)
+        bool deleteLastColumn)
     {
         using var xlWorkbook = new XLWorkbook();
 
         var xlWorksheet = xlWorkbook.AddWorksheet(sheetName);
-        xlWorksheet.Cell(row: 1, column: 1).InsertTable(records).Cells().Style
+        var xlTable = xlWorksheet.Cell(row: 1, column: 1).InsertTable(records);
+        if (deleteLastColumn) xlTable.Column(xlTable.Columns().Count()).Delete();
+        xlTable.Cells().Style
             .Alignment.SetWrapText(true)
             .Alignment.SetVertical(XLAlignmentVerticalValues.Top);
         xlWorksheet.Columns().AdjustToContents(minWidth: 8d, maxWidth: 80d);
-        if (hideLastColumn) xlWorksheet.Column(xlWorksheet.Columns().Count()).Hide();
 
         var memoryStream = new MemoryStream();
         xlWorkbook.SaveAs(memoryStream);
