+++ src/WebApp/Pages/Staff/Complaints/DownloadSearch.cshtml.cs
@@ -1,13 +1,13 @@
 ﻿using Cts.AppServices.Attachments;
-using Cts.AppServices.Complaints;
 using Cts.AppServices.Complaints.QueryDto;
+using Cts.AppServices.DataExport;
 using Cts.AppServices.Permissions;
 using Cts.AppServices.Utilities;
 
 namespace Cts.WebApp.Pages.Staff.Complaints;
 
 [Authorize(Policy = nameof(Policies.ActiveUser))]
-public class DownloadSearchModel(IComplaintService complaints, IAuthorizationService authorization) : PageModel
+public class DownloadSearchModel(IDataExportService dataExportService, IAuthorizationService authorization) : PageModel
 {
     public ComplaintSearchDto Spec { get; private set; } = default!;
     public int ResultsCount { get; private set; }
@@ -18,7 +18,7 @@ public class DownloadSearchModel(IComplaintService complaints, IAuthorizationSer
         spec.TrimAll();
         if (!(await authorization.AuthorizeAsync(User, nameof(Policies.DivisionManager))).Succeeded)
             spec.DeletedStatus = null;
-        ResultsCount = await complaints.CountAsync(spec, token);
+        ResultsCount = await dataExportService.CountAsync(spec, token);
         Spec = spec;
         return Page();
     }
@@ -29,8 +29,8 @@ public class DownloadSearchModel(IComplaintService complaints, IAuthorizationSer
         spec.TrimAll();
         if (!(await authorization.AuthorizeAsync(User, nameof(Policies.DivisionManager))).Succeeded)
             spec.DeletedStatus = null;
-        return File((await complaints.ExportSearchAsync(spec, token))
-            .AsExcelStream(sheetName: "CTS Search Results", hideLastColumn: spec.DeletedStatus == null),
+        return File((await dataExportService.ExportSearchAsync(spec, token))
+            .AsExcelStream(sheetName: "CTS Search Results", deleteLastColumn: spec.DeletedStatus == null),
             FileTypes.ExcelContentType, fileDownloadName: $"cts_search_{DateTime.Now:yyyy-MM-dd--HH-mm-ss}.xlsx");
     }
 }
